Initial Commit of Microsoft Edge WebGL Implementation (Transpiler Only)
This commit contains a select number of files from the implementation of WebGL that ships
in the Microsoft Edge browser. The source is provided together with tests, as a reference
for engineers who are creating web browsers and related applications.

By publishing this code, we hope to facilitate the investigation and resolution of any
interoperability issues that may exist between different implementations of the standard.

- Core, Foundation
Header files for common data structures used throughout the code

- GLSLparse
GLSL to HLSL transpiler code

- ft_glslparse
Test suite for the transpiler

- How do I build this?
At this time, we are publishing the source code for reference only. We do plan to provide
project files and instructions to generate binaries down the line, however we do not have
a target date for it just yet. If you have specific questions or needs, please do reach out:
we will be happy to evaluate your scenario and discuss how we can help.